chore(user route): remove user's id from url - remove user's id from url - group components into folders - add jsdoc comments
[Delivers #158556992]
