--- client/components/PageFooter.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const PageFooter = () => (
-    <footer className="page-footer foot lighten-5">
-        <div className="footer-copyright">
-            <div className="container center-align grey-text">
-                © 2017 Copyright Hello-books
-                    </div>
-        </div>
-    </footer>
-);
-
-export default PageFooter;
