+++ client/components/User/User.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
-import Navbar from '../Navbar';
-import PageFooter from '../PageFooter';
+import Navbar from '../Common/Navbar.jsx';
+import PageFooter from '../Common/PageFooter.jsx';
 import Favorites from './Favorites.jsx';
 import Profile from './Profile.jsx';
 import NotFound from '../Common/NotFound.jsx';
@@ -17,9 +17,9 @@ const User = ({ match }) => (
   <div>
         <Navbar />
         <Switch>
-          <Route exact path={`${match.url}/:userId/favorites`}
+          <Route exact path={`${match.url}/favorites`}
             component={ Favorites }/>
-          <Route exact path={`${match.url}/:userId/profile`}
+          <Route exact path={`${match.url}/profile`}
             component={ Profile }/>
           )}/>
           <Route component={ NotFound } />
